+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fixed accessing unmounted state on closing Assistant panel.
+- Showing athletics images [#5240](https://github.com/rokwire/illinois-app/issues/5240).
 
 ## [7.2.11] - 2025-07-15
 ### Changed